+++ tests/test_helpers.py
@@ -30,50 +30,56 @@ def app() -> Quart:
 
 @pytest.mark.asyncio
 async def test_make_response(app: Quart) -> None:
-    with app.app_context():
+    async with app.app_context():
         response = await make_response('foo', 202)
         assert response.status_code == 202
         assert b'foo' in (await response.get_data())
 
 
-def test_flash(app: Quart) -> None:
-    with app.test_request_context('GET', '/'):
-        flash('message')
+@pytest.mark.asyncio
+async def test_flash(app: Quart) -> None:
+    async with app.test_request_context('GET', '/'):
+        await flash('message')
         assert get_flashed_messages() == ['message']
         assert get_flashed_messages() == []
 
 
-def test_flash_category(app: Quart) -> None:
-    with app.test_request_context('GET', '/'):
-        flash('bar', 'error')
-        flash('foo', 'info')
+@pytest.mark.asyncio
+async def test_flash_category(app: Quart) -> None:
+    async with app.test_request_context('GET', '/'):
+        await flash('bar', 'error')
+        await flash('foo', 'info')
         assert get_flashed_messages(with_categories=True) == [('error', 'bar'), ('info', 'foo')]
         assert get_flashed_messages() == []
 
 
-def test_flash_category_filter(app: Quart) -> None:
-    with app.test_request_context('GET', '/'):
-        flash('bar', 'error')
-        flash('foo', 'info')
+@pytest.mark.asyncio
+async def test_flash_category_filter(app: Quart) -> None:
+    async with app.test_request_context('GET', '/'):
+        await flash('bar', 'error')
+        await flash('foo', 'info')
         assert get_flashed_messages(category_filter=['error']) == ['bar']
         assert get_flashed_messages() == []
 
 
-def test_url_for(app: Quart) -> None:
-    with app.app_context():
+@pytest.mark.asyncio
+async def test_url_for(app: Quart) -> None:
+    async with app.app_context():
         assert url_for('index') == '/'
         assert url_for('index_post', _method='POST') == '/post'
         assert url_for('resource', id=5) == '/resource/5'
 
 
-def test_url_for_external(app: Quart) -> None:
-    with app.app_context():
+@pytest.mark.asyncio
+async def test_url_for_external(app: Quart) -> None:
+    async with app.app_context():
         assert url_for('index', _external=True) == 'http://localhost/'
         assert url_for('resource', id=5, _external=True) == 'http://localhost/resource/5'
 
 
-def test_url_for_scheme(app: Quart) -> None:
-    with app.app_context():
+@pytest.mark.asyncio
+async def test_url_for_scheme(app: Quart) -> None:
+    async with app.app_context():
         with pytest.raises(ValueError):
             url_for('index', _scheme='https')
         assert url_for('index', _scheme='https', _external=True) == 'https://localhost/'
@@ -82,13 +88,15 @@ def test_url_for_scheme(app: Quart) -> None:
         ) == 'https://localhost/resource/5'
 
 
-def test_url_for_anchor(app: Quart) -> None:
-    with app.app_context():
+@pytest.mark.asyncio
+async def test_url_for_anchor(app: Quart) -> None:
+    async with app.app_context():
         assert url_for('index', _anchor='&foo') == '/#%26foo'
         assert url_for('resource', id=5, _anchor='&foo') == '/resource/5#%26foo'
 
 
-def test_url_for_blueprint_relative(app: Quart) -> None:
+@pytest.mark.asyncio
+async def test_url_for_blueprint_relative(app: Quart) -> None:
     blueprint = Blueprint('blueprint', __name__)
 
     @blueprint.route('/')
@@ -97,6 +105,6 @@ def test_url_for_blueprint_relative(app: Quart) -> None:
 
     app.register_blueprint(blueprint, url_prefix='/blue')
 
-    with app.test_request_context('GET', '/blue/'):
+    async with app.test_request_context('GET', '/blue/'):
         assert url_for('.index') == '/blue/'
         assert url_for('index') == '/'